requirements.txt
discord.py
google-api-python-client
google-auth
gspread
oauth2client